Make the initial commit of libgit-rs
libgit-rs is a pure rust library to interact with git repositories on
disk, much like libgit2. libgit-rs is meant to provide an idiomatic API
for interacting with a git repository and it's contents whether by
building a CLI tool, a git subcommand or just to use in a Rust program
without needing to bind to any C dependencies beyond the usual OS libs
to interact with the filesystem. It's goal is to be well tested, easy to
maintain and use with a preference for ease of use over maintenance, and
to provide well documented examples and API documentation.

This commit starts this library by creating a `Blob` type to represent a
`Blob` in the git object database and an `OID` to represent the Sha1
hash of these objects in the database.